added permission from update delete user and tests

+++ app/services/user_service.py
@@ -5,10 +5,6 @@ import bcrypt
 from loguru import logger
 from sqlalchemy.ext.asyncio import AsyncSession
 
-from app.conf.detail import Messages
-from app.exept.custom_exceptions import UserNotFound, NotFound
-from app.schemas.users import UserSchema, UserUpdateRequest, BaseUserSchema
-from app.repository.user_repository import UserRepository
 from app.repository.user_repository import UserRepository
 from app.schemas.users import UserSchema, UserUpdateRequest, BaseUserSchema
 from app.conf.detail import Messages
@@ -17,13 +13,21 @@ from app.exept.custom_exceptions import (
     EmailAlreadyExists,
     UserAlreadyExists,
     NotFound,
+    NotPermission,
 )
+from app.utils import password_utils
 
 
 class UserService:
     def __init__(self, session: AsyncSession, repository: UserRepository):
         self.session = session
         self.repository = repository
+
+    @staticmethod
+    async def check_user_permission(user_id: uuid.UUID, current_user: UserSchema):
+        if user_id != current_user.id:
+            logger.info(Messages.NOT_PERMISSION)
+            raise NotPermission()
 
     async def _get_user_or_raise(self, user_id: uuid.UUID) -> UserSchema:
         user = await self.repository.get_one(id=user_id)
@@ -32,7 +36,7 @@ class UserService:
             raise UserNotFound()
         logger.info(Messages.SUCCESS_GET_USER)
 
-        return UserSchema.from_orm(user)
+        return UserSchema.model_validate(user)
 
     async def create_user(self, data: dict) -> UserSchema:
         email = data.get("email")
@@ -61,8 +65,8 @@ class UserService:
 
         user = await self.repository.create_one(user_data)
         logger.info(Messages.SUCCESS_CREATE_USER)
-        
-        return UserSchema.from_orm(user)
+
+        return UserSchema.model_validate(user)
 
     async def get_users(self, skip: int = 1, limit: int = 10) -> List[UserSchema]:
         users = await self.repository.get_many(skip=skip, limit=limit)
@@ -71,27 +75,38 @@ class UserService:
             raise NotFound()
         logger.info(Messages.SUCCESS_GET_USERS)
 
-        return users
+        return [UserSchema.model_validate(user) for user in users]
 
     async def get_user_by_id(self, user_id: uuid.UUID) -> Optional[UserSchema]:
         return await self._get_user_or_raise(user_id)
 
     async def update_user(
-        self, user_id: uuid.UUID, update_data: UserUpdateRequest
+        self,
+        user_id: uuid.UUID,
+        update_data: UserUpdateRequest,
+        current_user: UserSchema,
     ) -> UserSchema:
+        await self.check_user_permission(user_id, current_user)
         await self._get_user_or_raise(user_id)
-        update_dict = update_data.dict(exclude_unset=True)
+        existing_user = await self.repository.get_one(username=update_data.username)
+        if existing_user:
+            logger.info(Messages.USER_ALREADY_EXISTS)
+            raise UserAlreadyExists()
 
-        if "password" in update_dict:
-            update_dict["password"] = bcrypt.hashpw(
-                update_dict["password"].encode("utf-8"), bcrypt.gensalt()
-            ).decode("utf-8")
+        update_dict = update_data.model_dump(exclude_unset=True)
+
+        if update_data.password:
+            hashed_password = password_utils.hash_password(update_dict["password"])
+            update_dict["password"] = hashed_password.decode("utf-8")
 
         updated_user = await self.repository.update_one(user_id, update_dict)
         logger.info(Messages.SUCCESS_UPDATE_USER)
-        return UserSchema.from_orm(updated_user)
+        return UserSchema.model_validate(updated_user)
 
-    async def delete_user(self, user_id: uuid.UUID) -> BaseUserSchema:
+    async def delete_user(
+        self, user_id: uuid.UUID, current_user: UserSchema
+    ) -> BaseUserSchema:
+        await self.check_user_permission(user_id, current_user)
         await self._get_user_or_raise(user_id)
         logger.info(Messages.SUCCESS_DELETE_USER)
         return await self.repository.delete_one(user_id)
